backend/requirements.txt
fastapi==0.121.1
pymysql==1.1.2
python-dotenv==1.2.1
starlette==0.49.3
uvicorn[standard]==0.38.0
